services/kb-ocr-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pillow==10.3.0
pytesseract==0.3.10
